Stop using MainAreaWidget
Using MainAreaWidget gave us a Lumino toolbar. However, I
want this to be as pure react as possible - do not want to learn
an entire new system of message passing (Signals) and try hook
that up to react's.

Instead, we use a Panel, which mostly leaves us alone. We just
render our ReactWidget into that. Needs to be styled separately,
but that's ok.

- Remove redundant divs, using React fragments instead
- Use classes everywhere, avoid ids.

Co-authored-by: tanvibhakta <tanvibhakta@gmail.com>