Fetch list of rooms from serverextension
Instead of hard-coding the list of rooms in the client,
we hardcode it in the serverextension instead. We make a
request on component mount to get list of rooms.

Eventually, there will probably be a hub-wide service
that determines who gets what rooms. However, by talking
to it via the serverextension rather than directly, we can
do local development very easily. It also means the service
itself is optional, which is important for smaller installations.

Woop! Woop!